backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
openai==1.59.7
python-dotenv==1.0.1
httpx==0.28.1
